moved toDatetimeLocalString to a more suitable location
