+++ src/components/Forms/prepopulated-event-form.tsx
@@ -5,8 +5,8 @@ import Button from "@/components/Button";
 import FieldInput from "@/components/FormFields/field-input";
 import FieldSingleSelect from "@/components/FormFields/field-single-select";
 import FieldTextArea from "@/components/FormFields/field-text-area";
+import { toDatetimeLocalString } from "@/components/utils/format-date";
 import { ItemType, useItemTypes } from "@/hooks/item-types";
-import { toDatetimeLocalString } from "@/lib/date-utils";
 import { updateEventSchema } from "@/schema/event";
 import { Event } from "@/types";
 import { UpdateEvent } from "@/types";
